backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
requests==2.32.3
python-dotenv==1.0.1
APScheduler==3.10.4
python-dateutil==2.9.0.post0
aiofiles==23.2.1
